+++ src/screens/home/index.ts
@@ -1,2 +1,3 @@
 // Home Screens
-// Example: export { HomeScreen } from './HomeScreen';
+export { default as HomeScreen } from './HomeScreen';
+export { default } from './HomeScreen';